backend/scrumtool/requirements.txt
alembic==1.4.2
asgiref==3.2.3
astroid==2.3.3
Babel==2.8.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.8
Django==3.0.3
django-crispy-forms==1.9.0
django-polymorphic==2.1.2
django-rest-auth==0.9.5
djangoajax==3.2
djangorestframework==3.11.0
docutils==0.16
Flask==1.0.2
Flask-BabelEx==0.9.4
Flask-Compress==1.4.0
Flask-Gravatar==0.5.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
paramiko==2.7.1
passlib==1.7.1
Pillow==7.1.1
psutil==5.5.1
psycopg2-binary==2.8.4
pyaml==20.3.1
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
PyNaCl==1.3.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2018.9
PyYAML==5.3.1
requests==2.23.0
rope==0.16.0
simplejson==3.16.0
six==1.14.0
speaklater==1.3
SQLAlchemy==1.3.15
sqlparse==0.2.4
sshtunnel==0.1.5
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
django-cors-headers==3.2.1
django-oauth-toolkit==1.3.2
